Added pydantic_validators utils
